% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw.f.BGMboot2.R
\name{bw.f.BGMboot2}
\alias{bw.f.BGMboot2}
\title{\insertCite{borrajo2017;textual}{WData} bootstrap bandwidth selection for \insertCite{jones1991;textual}{WData} kernel density estimator}
\usage{
bw.f.BGMboot2(
  y,
  w = function(y) {
     ifelse(y >= 0, y, NA)
 },
  kernel = c("gaussian", "epanechnikov", "rectangular", "triangular", "biweight",
    "cosine", "optcosine"),
  bw0 = 1/8 * n^(-1/9),
  lower = IQR(y) * n^(-0.2) * 2000^(-1),
  upper = IQR(y) * (log(n)/n)^(0.2) * 500,
  nh = 200L,
  tol = 0.1 * lower,
  from = min(y) - (sort(y)[5] - min(y)),
  to = max(y) + (max(y) - sort(y, decreasing = T)[5]),
  plot = TRUE
)
}
\arguments{
\item{y}{A numeric vector containing the biased data.}

\item{w}{A function representing the bias function applied to the data points. It must be evaluable and positive in each point of \code{y}. By default, it is set to the length-biased function.}

\item{kernel}{A character vector specifying the kernel function. Available options: \code{"gaussian"}, \code{"epanechnikov"}, \code{"rectangular"}, \code{"triangular"}, \code{"biweight"}, \code{"cosine"} and \code{"optcosine"}.}

\item{bw0}{A character string specifying the pilot bandwidth. It can also be a character string giving a rule to choose the bandwidth. Options available can be checked in \code{\link[stats:bandwidth]{bw.nrd}}. By default it is set to \verb{1/8 n^(-1/9)}.}

\item{lower}{Numeric value specifying the lower bound for bandwidth selection. Default is computed based on the interquartile range (IQR) and number of data points.}

\item{upper}{Numeric value specifying the upper bound for bandwidth selection. Default is computed based on the interquartile range (IQR) and number of data points.}

\item{nh}{An integer specifying the number of points for evaluating the MISE function. Default is 200.}

\item{tol}{Tolerance value used for checking if the minimum bandwidth occurs at the boundary of the range. Default is 10\% of the lower bound.}

\item{from}{Numeric value specifying the lower bound for  evaluation. Default is calculated based on the range of input data.}

\item{to}{Numeric value specifying the upper bound for  evaluation. Default is calculated based on the range of input data.}

\item{plot}{Logical value indicating whether to plot the function to be minimized. Default is \code{TRUE}.}
}
\value{
The estimated bootstrap bandwidth for \insertCite{jones1991;textual}{WData} kernel density estimator.
}
\description{
This function estimates the bandwidth for \insertCite{jones1991;textual}{WData} kernel density estimator using the bias-corrected bootstrap method developed by \insertCite{borrajo2017;textual}{WData} and based on the methodology introduced by \insertCite{bose2013;textual}{WData}.
}
\details{
In \insertCite{bose2013;textual}{WData}, a bootstrap method is provided for bandwidth selection in the context of unbiased data that does not require assuming any specific form for the density function.
The adaptation of this method to the context of biased data, carried out by \insertCite{borrajo2017;textual}{WData}, is presented below.
\itemize{
\item Given a biased sample \eqn{\{Y_1, Y_2, \dots, Y_n\}}, the estimator \eqn{\widehat{g}_{n, h_0}} with pilot bandwidth \eqn{h_{0}} is constructed as
\deqn{\widehat{g}_{n, h_0}(y)= \frac{1}{n}\sum_{i=1}^{n} L_{h_0}(y-Y_i).}
\item A bootstrap sample \eqn{\{Y_1^{\ast}, Y_2^{\ast}, \dots, Y_n^{\ast}\}} is generated with density function \eqn{\widehat{g}_{n, h_0}(y)}.
\item From \eqn{\{Y_1^{\ast}, Y_2^{\ast}, \dots, Y_n^{\ast}\}}, an unbiased sample of the variable \eqn{Y^{\ast}} is generated, and the bootstrap density estimator \eqn{\widehat{f}^{\ast}_{\mathrm{J}, h}(y)} with bandwidth \eqn{h} is constructed as
\deqn{\widehat{f}^{\ast}_{\mathrm{J}, h}(y)= \frac{\widehat{\mu}_{w}^{\ast}}{n}\sum_{i=1}^{n} \frac{1}{w(Y_{i}^{\ast})} K^{\ast}_h(y-Y_{i}^{\ast}),
\quad \text{where} \quad
\widehat{\mu}_{w}^{\ast}= n\left(\sum_{i=1}^{n} \frac{1}{w(Y_{i}^{\ast})} \right)^{-1}.}
}

Under certain conditions, it is satisfied that:
\deqn{\begin{aligned}
\mathrm{MSE}^{\ast}\left(\widehat{f}^{\ast}_{\mathrm{J}, h}(y)\right) & =
\left(\frac{\int_{-\infty}^{+ \infty} w(z)^{-1} K^{\ast}_{h}(y-z) \widehat{g}_{n, h_0}(z) d z}
{\int_{-\infty}^{+ \infty} w(z)^{-1} \widehat{g}_{n, h_0}(z) d z}
-\widehat{f}_{\mathrm{J},h}(y)\right)^2
+\frac{1}{n\left(\int_{-\infty}^{+ \infty} w(z)^{-1} \widehat{g}_{n, h_0}(z) d z\right)^2} \\ &
\left[ \int_{-\infty}^{+ \infty} w(z)^{-2} \left(K^{\ast}_{h}(y-z)\right)^2 \widehat{g}_{n, h_0}(z) d z
-\left(\int_{-\infty}^{+ \infty} w(z)^{-1} K^{\ast}_{h}(y-z) \widehat{g}_{n, h_0}(z) d z \right)^2\right] \\ &
+O\left(\frac{1}{n}\right).
\end{aligned}}

Hence, it is possible to obtain a bandwidth \eqn{\widehat{h}_{\mathrm{B}}} minimizing \eqn{\mathrm{MISE}^{\ast}(h)} over a compact interval \eqn{[h_1,h_2]}, i.e.,
\deqn{
\widehat{h}_{\mathrm{B}} = \arg \min_{h \in [h_1,h_2]} \mathrm{MISE}^{\ast}\left(\widehat{f}^{\ast}_{J, h}\right) =
\arg \min_{h \in [h_1,h_2]} \int_{-\infty}^{+ \infty} \mathrm{MSE}^{\ast}\left(\widehat{f}^{\ast}_{J, h}(y)\right)dy.
}
}
\examples{
bw.f.BGMboot2(shrub.data$Width, nh = 50L)
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=df.jones]{df.jones}}
}
