% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw.f.BGMcv.R
\name{bw.f.BGMcv}
\alias{bw.f.BGMcv}
\title{Cross-validation bandwidth selection for \insertCite{jones1991;textual}{WData} kernel density estimator}
\usage{
bw.f.BGMcv(
  y,
  w = function(y) {
     ifelse(y >= 0, y, NA)
 },
  kernel = c("gaussian", "epanechnikov", "rectangular", "triangular", "biweight",
    "cosine", "optcosine"),
  lower = IQR(y) * n^(-0.2) * 2000^(-1),
  upper = IQR(y) * (log(n))^(0.2) * n^(-0.2) * 500,
  nh = 200L,
  tol = 0.1 * lower,
  plot = TRUE
)
}
\arguments{
\item{y}{A numeric vector containing the biased data.}

\item{w}{A function representing the bias function applied to the data points. It must be evaluable and positive in each point of \code{y}. By default, it is set to the length-biased function.}

\item{kernel}{A character vector specifying the kernel function. Available options: \code{"gaussian"}, \code{"epanechnikov"}, \code{"rectangular"}, \code{"triangular"}, \code{"biweight"}, \code{"cosine"} and \code{"optcosine"}.}

\item{lower}{Numeric value specifying the lower bound for bandwidth selection. Default is computed based on the interquartile range (IQR) and number of data points.}

\item{upper}{Numeric value specifying the upper bound for bandwidth selection. Default is computed based on the interquartile range (IQR) and number of data points.}

\item{nh}{An integer specifying the number of points for evaluating the cross-validation function. Default is 200.}

\item{tol}{Tolerance value used for checking if the minimum bandwidth occurs at the boundary of the range. Default is 10\% of the lower bound.}

\item{plot}{Logical value indicating whether to plot the cross-validation function. Default is \code{TRUE}.}
}
\value{
The optimal cross-validation bandwidth for \insertCite{jones1991;textual}{WData} kernel density estimator.
}
\description{
This function performs bandwidth selection for \insertCite{jones1991;textual}{WData} kernel density estimation using cross-validation. It iterates through a range of bandwidth values and computes the cross-validation score for each bandwidth. The bandwidth that minimizes the cross-validation score is selected as the optimal bandwidth.
}
\details{
The optimal bandwidth is the one that minimizes the cross-validation function, i.e.,
\deqn{
\widehat{h}_{\mathrm{CV}}= \mathrm{arg} \min_{h>0} \mathrm{CV}(h) = \mathrm{arg} \min_{h>0} \int_{-\infty}^{+\infty} \widehat{f}_{\mathrm{J}}(y)^2 d y-2 \widehat{\mathbb{E}[\widehat{f}_{\mathrm{J}}]}.}
It holds that
\deqn{
\int_{-\infty}^{+\infty} \widehat{f}_{\mathrm{J}}(y)^2 dy = \frac{\widehat{\mu}_w^2}{n^{2} h} \sum_{i=1}^{n} \sum_{\substack{j=1}}^{n}
\frac{1}{w(Y_i)} \frac{1}{w(Y_j)}
(K \circ K) \left(\frac{Y_i-Y_j}{h} \right),
}
where \eqn{\circ} denotes convolution between two functions. Moreover,
\deqn{
\widehat{\mathbb{E}[\widehat{f}_{\mathrm{J}}]}=\frac{\widehat{\mu}_w}{n} \sum_{i=1}^n \frac{1}{w(Y_i)} \widehat{f}_{\mathrm{J}, -i}\left(Y_i\right),
\quad \text{where} \quad
\widehat{\mu}_w=n \left(\sum_{i=1}^{n}  \frac{1}{w(Y_i)} \right)^{-1}
}
and \eqn{\widehat{f}_{J, -i}} is \insertCite{jones1991;textual}{WData} estimator constructed without the \eqn{i}-th data point. In practice, it is not necessary to calculate \eqn{\widehat{f}_{J, -i}} for each data point in the sample, but the estimator
\eqn{\widehat{\mathbb{E}[\widehat{f}_{\mathrm{J}}]}} can be calculated as
\deqn{
\widehat{\mathbb{E}[\widehat{f}_{\mathrm{J}}]} = \frac{\widehat{\mu}_w}{n} \sum_{i=1}^{n}  \frac{1}{w(Y_i)} \left( \sum_{j\neq i}  \frac{1}{w(Y_j)} \right)^{-1} \left(\sum_{j\neq i}  \frac{1}{w(Y_j)} K_h(Y_i-Y_j)\right).
}
The function optimizes the cross-validation function on the interval \eqn{I} determined by \code{lower} and \code{upper}. By default,
\eqn{I} is the one suggested by \insertCite{borrajo2017;textual}{WData}:
\deqn{
I = \left[\frac{\mathrm{IQR}}{2000n^{1/5}}, \frac{500 \mathrm{IQR} \log(n)^{1/5}}{n^{1/5}}\right],
}
where IQR is the interquartile range.
}
\examples{
bw.f.BGMcv(shrub.data$Width)
bw.f.BGMcv(shrub.data$Width, kernel = "epanechnikov")

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=df.jones]{df.jones}}
}
