% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw.F.SBCnrd0.R
\name{bw.F.SBCnrd0}
\alias{bw.F.SBCnrd0}
\title{Rule of the thumb bandwidth selection for \insertCite{bose2022;textual}{WData} kernel distribution estimator}
\usage{
bw.F.SBCnrd0(
  y,
  w = function(y) {
     ifelse(y >= 0, y, NA)
 },
  kernel = c("gaussian", "epanechnikov", "rectangular", "triangular", "biweight",
    "cosine", "optcosine")
)
}
\arguments{
\item{y}{A numeric vector containing the biased data.}

\item{w}{A function representing the bias function applied to the data points. It must be evaluable and positive in each point of \code{y}. By default, it is set to the length-biased function.}

\item{kernel}{A character vector specifying the kernel function. Available options: \code{"gaussian"}, \code{"epanechnikov"}, \code{"rectangular"}, \code{"triangular"}, \code{"biweight"}, \code{"cosine"} and \code{"optcosine"}.}
}
\value{
The optimal bandwidth for \insertCite{bose2022;textual}{WData} kernel distribution estimator based on the rule of the thumb.
}
\description{
This function calculates the bandwidth for  \insertCite{bose2022;textual}{WData} kernel distribution estimator using the rule of the thumb.
}
\details{
The bandwidth is computed as follows:
\deqn{\widehat{h}_{F, \mathrm{RT}} =
\left(\frac{4 \sqrt{\pi} \widehat{\mu}_{w} \widehat{\bar{\mu}}_{w}\kappa(K)}{n \eta(K)^{2} }\right)^{1/3} \widehat{\sigma}, }
where both \eqn{\kappa(K)} and \eqn{\eta(K)} depend only on the kernel,
\deqn{\widehat{\mu}_w=n \left(\sum_{i=1}^{n}  \frac{1}{w(Y_i)} \right)^{-1}
\quad \text{and} \quad
\widehat{\bar{\mu}}_w= \frac{\widehat{\mu}_w}{n} \sum_{i=1}^{n} \frac{1}{w(Y_i)^2}.}
\eqn{\widehat{\sigma}} is an estimate of the standard deviation of the distribution given by
\deqn{\widehat{\sigma}_w=\sqrt{\left(\frac{1}{n} \sum_{i=1}^n \frac{1}{w(Y_i)}\right)^{-1}
\left[\left(\frac{1}{n} \sum_{i=1}^n w(Y_i)\right)-\left(\frac{1}{n} \sum_{i=1}^n \frac{1}{w(Y_i)}\right)^{-1}\right]}.}
}
\examples{
bw.F.SBCnrd0(shrub.data$Width)
bw.F.SBCnrd0(shrub.data$Width, kernel = "epanechnikov")

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=cdf.bd]{cdf.bd}}
}
