% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw.F.SBCpi.R
\name{bw.F.SBCpi}
\alias{bw.F.SBCpi}
\title{Plug-in bandwidth selection for \insertCite{bose2022;textual}{WData} kernel distribution estimator}
\usage{
bw.F.SBCpi(
  y,
  w = function(y) {
     ifelse(y >= 0, y, NA)
 },
  kernel = c("gaussian", "epanechnikov", "rectangular", "triangular", "biweight",
    "cosine", "optcosine")
)
}
\arguments{
\item{y}{A numeric vector for which the density estimation is performed.}

\item{w}{A function representing the bias function to be used in the estimation. It must be evaluable and positive in each point of \code{y}. By default, it is set to the length-biased function.}

\item{kernel}{A character vector specifying the kernel function. Available options: \code{"gaussian"}, \code{"epanechnikov"}, \code{"rectangular"}, \code{"triangular"}, \code{"biweight"}, \code{"cosine"} and \code{"optcosine"}.}
}
\value{
The optimal bandwidth for \insertCite{bose2022;textual}{WData} kernel distribution estimator using the plug-in method.
}
\description{
This function calculates the bandwidth for \insertCite{bose2022;textual}{WData} kernel distribution estimator using the plug-in method.
}
\details{
The bandwidth is given by:
\deqn{\widehat{h}_{\mathrm{PI}} = \left(\frac{\widehat{\mu}_w \widehat{\bar{\mu}}_w \tau_{W}^{2} }{n \sigma_K^4 R \left(\widehat{f}_{\mathrm{J},\widehat{g}}^{(1)}\right)}\right)^{1/3},
\quad \text{where} \quad
\widehat{\mu}_w=n \left(\sum_{i=1}^{n}  \frac{1}{w(Y_i)} \right)^{-1},
\quad
\widehat{\bar{\mu}}_w= \frac{\widehat{\mu}_w}{n} \sum_{i=1}^{n} \frac{1}{w(Y_i)^2}}
and \eqn{\tau_{W}^{2}} depends only on the kernel function. \eqn{\widehat{g}} is an estimator of:
\deqn{g = \arg \min_{g>0} \mathrm{AMSE} \left(R\left(\widehat{f}_{\mathrm{J},g}^{(1)}\right)\right)
= \left(\frac{3 \mu_w \bar{\mu}_w R\left(L^{(1)}\right)}{2 n \sigma_L^2 R\left( f^{(2)} \right)^2}\right)^{1/5}.}
Here, \eqn{R\left(f^{(2)}\right)} is estimated assuming that \eqn{f} follows a normal distribution.
}
\examples{
bw.F.SBCpi(shrub.data$Width, kernel = "epanechnikov")
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=cdf.bd]{cdf.bd}}
}
