% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw.q.SBCcv.R
\name{bw.q.SBCcv}
\alias{bw.q.SBCcv}
\title{Cross-validation bandwidth selection for kernel quantile estimator}
\usage{
bw.q.SBCcv(
  y,
  w = function(y) {
     ifelse(y >= 0, y, NA)
 },
  kernel = c("gaussian", "epanechnikov", "rectangular", "triangular", "biweight",
    "cosine", "optcosine"),
  lower = 1e-04,
  upper = 0.2,
  nh = 200L,
  tol = 0.1 * lower,
  plot = TRUE
)
}
\arguments{
\item{y}{A numeric vector containing the biased data.}

\item{w}{A function representing the bias function applied to the data points. It must be evaluable and positive in each point of \code{y}. By default, it is set to the length-biased function.}

\item{kernel}{A character vector specifying the kernel function. Available options: \code{"gaussian"}, \code{"epanechnikov"}, \code{"rectangular"}, \code{"triangular"}, \code{"biweight"}, \code{"cosine"} and \code{"optcosine"}.}

\item{lower}{Numeric value specifying the lower bound for bandwidth selection. Default is 0.0001.}

\item{upper}{Numeric value specifying the upper bound for bandwidth selection. Default is 0.2.}

\item{nh}{An integer specifying the number of points for evaluating the cross-validation function. Default is 200.}

\item{tol}{Tolerance value used for checking if the minimum bandwidth occurs at the boundary of the range. Default is 10\% of the lower bound.}

\item{plot}{Logical value indicating whether to plot the cross-validation function. Default is \code{TRUE}.}
}
\value{
The optimal cross-validation bandwidth for kernel quantile estimator.
}
\description{
This function performs bandwidth selection for kernel quantile estimation using cross-validation. It iterates through a range of bandwidth values and computes the cross-validation score for each bandwidth. The bandwidth that minimizes the cross-validation score is selected as the optimal bandwidth.
}
\details{
The optimal bandwidth is the one that minimizes the cross-validation function. Specifically, the bandwidth \eqn{\widehat{h}_{\operatorname{CV}}} is defined as
\deqn{
  \widehat{h}_{\operatorname{CV}} = \arg \min_{h \in [h_1, h_2]} \frac{1}{n} \sum_{i=1}^{n} \left(Y_{(i)} - \widehat{F_{h,-i}^{-1}}(T_i)\right)^2,
}
where \eqn{\widehat{F_{h,-i}^{-1}}} denotes the quantile function estimator computed without the \eqn{i}-th order statistic.
}
\examples{
bw.q.SBCcv(shrub.data$Width)
bw.q.SBCcv(shrub.data$Width, kernel = "epanechnikov")

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=qf.SBCkernel]{qf.SBCkernel}}
}
