% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf.efro.R
\name{cdf.efro}
\alias{cdf.efro}
\title{\insertCite{efromovich2008;textual}{WData} adaptative density and distribution estimator}
\usage{
cdf.efro(
  y,
  w = function(y) {
     ifelse(y >= 0, y, NA)
 },
  x,
  from,
  to,
  nb = 512L,
  Jn = floor(4 + 0.5 * log(n)),
  Ct = 4,
  Cjm = 6,
  plot = TRUE
)
}
\arguments{
\item{y}{A numeric vector containing the biased data.}

\item{w}{A function representing the bias function applied to the data points. It must be evaluable and positive in each point of \code{y}. By default, it is set to the length-biased function.}

\item{x}{A numeric vector specifying the points where the distribution is estimated. Alternatively, \code{from}, \code{to} and \code{nb} can be used to define the evaluation points.}

\item{from}{Numeric value specifying the lower bound for  evaluation when \code{x} not provided. Default is calculated based on the range of input data.}

\item{to}{Numeric value specifying the upper bound for  evaluation when \code{x} not provided. Default is calculated based on the range of input data.}

\item{nb}{An integer specifying the number of points for evaluation when \code{x} not provided. Default is 512.}

\item{Jn}{Parameter which controls the maximum number of coefficients to be included in the low frequency part of the estimator.}

\item{Ct}{Parameters which controls the minimum size that the coefficients need to have to be included in the high frequency part of the estimation.}

\item{Cjm}{Parameter which controls the maximum number of coefficients to be included in the high frequency part of the estimator.}

\item{plot}{Logical value indicating whether to plot the estimations. Default is \code{TRUE}.}
}
\value{
A list with the following components:
\item{J}{The optimal cutoff \eqn{\widehat{J}}.}
\item{c}{The value for constant \eqn{c}.}
\item{U}{The value for the cutoff \eqn{U}.}
\item{x}{The points where the density is estimated.}
\item{fords}{The estimated density values.}
\item{Fords}{The estimated distribution values.}
\item{theta}{The estimated coefficients \eqn{\widehat{\theta_j}}.}
\item{omega}{The estimated weights \eqn{\widehat{\omega_j}}.}
\item{MISE}{MISE per each cutoff between zero and  \code{Cjm * Jn}.}
\item{x01}{The points in \eqn{[0,1]} where the density is estimated.}
\item{y01}{The estimator \eqn{\widehat{f}_{S, w, HF}(y)} evaluated in \code{x01}.}
}
\description{
This function computes \insertCite{efromovich2008;textual}{WData} density and distribution estimator given a biased dataset and its bias function.
}
\details{
Given a biased sample from a variable with support on \eqn{[0,1]},
the estimator proposed by \insertCite{efromovich2008;textual}{WData} is given by
\deqn{
\widehat{f}_{S, w, HF,c}(y) = \max \left( 0,  \widehat{f}_{S, w, HF}(y) - c\right),
}
where \eqn{c} is a constant ensuring that \eqn{\widehat{f}_{S, w, HF,c}}
integrates to \eqn{1} over \eqn{[0,1]}. Specifically,
\deqn{
\widehat{f}_{S, w, HF}(y) = \sum_{j=0}^{\widehat{J}} \widehat{\omega}_j \widehat{\theta}_j \varphi_j(y) + \sum_{j = \widehat{J}+1}^{c_{JM}J_n} \mathbb{I}\left(\widehat{\theta}_{j}^2 > c_T  \ln(n)/n\right) \widehat{\theta}_j \varphi _j (y)
}
where \eqn{\varphi_0(u)=1} and \eqn{\varphi_j(u)=\sqrt{2}\cos (\pi ju)} if
\eqn{j>0}. The first and second summations correspond to the low and high
frequency parts of the estimator, respectively. The parameters \eqn{c_{JM}}
and \eqn{J_n} control the maximum number of coefficients that constitute the
estimator \eqn{\widehat{f}_{S, w, HF}}. \insertCite{efromovich2008;textual}{WData}
suggests taking \eqn{c_{JM} = 6} and \eqn{c_{T}=4}. The weights
\eqn{\omega_j} are given by
\deqn{
\widehat{\omega}_{0} = 1
\quad \text{and} \quad
\widehat{\omega}_j = \max\left(0, 1 - (n \widehat{\theta}_j^2)^{-1}\right)
\quad j > 0
}
and the coefficients by
\deqn{
\widehat{\theta}_j =\widehat{\mu}_w n^{-1}\sum_{i=1}^{n} w^{-1}\left(Y_i\right) \varphi_j\left(Y_i\right) .
}
The threshold \eqn{\widehat{J}} is chosen as
\deqn{
\widehat{J} = \arg \min_{J \in [0,J_n]} \sum_{j=0}^{J}\left(2 n^{-1} - \widehat{\theta}_j^2 \right).
}
\insertCite{efromovich2008;textual}{WData} establishes \eqn{J_n = \lfloor 4 + 0.5
\ln{n} \rfloor} as a reasonable upper bound, where \eqn{\lfloor z \rfloor}
denotes the greatest integer less than or equal to \eqn{z}.

The results presented for the interval \eqn{[0,1]} can be readily applied to
a compact interval \eqn{[a,b]}. For this purpose, given a sample \eqn{\{Y_1,
Y_2, \dots Y_n\}}, consider the transformation \eqn{Y_{i, [0,1]} = (Y_i -
a)/(b-a) \in [0,1]} and \eqn{w(x)_{[0,1]} = w((b-a)x+a)}, from which
\eqn{\widehat{f}_{S, w, HF,c, [0,1]}(x)} would be computed as indicated.
Finally, the estimation of the density function for the variable \eqn{X} is
given by
\deqn{
\widehat{f}_{S, w, HF, c, [a,b]}(y) = (b-a)^{-1} \widehat{f}_{S, w, HF, c, [0,1]}\left(\frac{y-a}{b-a}\right)
}
and its integral is taken as the estimation for the distribution function.
In many practical applications, the interval \eqn{[a,b]} is unknown. In such
cases, it is considered and interval of the form \eqn{[Y_{(1)} - \delta_1, Y_{(n)} +
\delta_2]}, where \eqn{Y_{(1)} \leq Y_{(2)} \leq \dots \leq Y_{(n)}} is the
ordered sample, and
\deqn{
\widehat{\delta_1} = Y_{(2)}- Y_{(1)}
\quad
\quad
\widehat{\delta_2} = Y_{(n)}- Y_{(n-1)}.
}
}
\examples{
cdf.efro(y = shrub.data$Width)
}
\references{
\insertAllCited{}
}
