% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqf.R
\name{eqf}
\alias{eqf}
\alias{plot.eqf}
\alias{print.eqf}
\alias{summary.eqf}
\alias{quantile.eqf}
\title{Create an "eqf" structure}
\usage{
eqf(x, probs = seq(0, 1, 0.25))

\method{plot}{eqf}(
  x,
  ...,
  ylab = "",
  xlab = "",
  verticals = FALSE,
  col.01line = "gray70",
  pch = 19
)

\method{print}{eqf}(x, digits = getOption("digits") - 2L, ...)

\method{summary}{eqf}(object, ...)

\method{quantile}{eqf}(x, ...)
}
\arguments{
\item{x}{A numeric vector or other data type that serves as the basis for the structure.}

\item{probs}{A numeric vector of probabilities (default is \code{seq(0, 1, 0.25)}).}

\item{...}{Additional arguments to be passed to methods.}

\item{xlab, ylab}{A character string for the x-axis/y-axis label for method \code{\link[graphics:plot.default]{plot}}.}

\item{verticals}{A logical value indicating whether to add vertical lines to the plot (default is \code{FALSE}).}

\item{col.01line}{A character string for the color of the 0-1 lines (default is \code{"gray70"}).}

\item{pch}{An integer specifying the plotting character for method \code{\link[graphics:plot.default]{plot}} (default is 19).}

\item{digits}{An integer specifying the number of significant digits to be printed (default is \code{getOption("digits") - 2L}).}

\item{object}{An object of class \code{"eqf"}.}
}
\value{
An object of class \code{"eqf"}, which is a list with elements \code{x} and \code{probs}.
}
\description{
This function generates an object of class \code{"eqf"} containing a vector and its associated probabilities.
}
\examples{
eqf(1:5, (1:5) / 5)
}
\seealso{
\code{\link[=qf.sen]{qf.sen}}, \code{\link[=qf.SBC]{qf.SBC}}
}
