% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qf.SBCkernel.R
\name{qf.SBCkernel}
\alias{qf.SBCkernel}
\title{Kernel quantile function estimator}
\usage{
qf.SBCkernel(
  y,
  w = function(y) {
     ifelse(y >= 0, y, NA)
 },
  x,
  bw = "bw.q.SBCcv",
  adjust = 1,
  kernel = c("gaussian", "epanechnikov", "rectangular", "triangular", "biweight",
    "cosine", "optcosine"),
  from = 0.01,
  to = 0.99,
  nb = 99,
  plot = TRUE,
  ...
)
}
\arguments{
\item{y}{A numeric vector containing the biased data.}

\item{w}{A function representing the bias function applied to the data points. It must be evaluable and positive in each point of \code{y}. By default, it is set to the length-biased function.}

\item{x}{A numeric vector specifying the points where the quantile is estimated. Alternatively, \code{from}, \code{to} and \code{nb} can be used to define the evaluation points.}

\item{bw}{The smoothing bandwidth to be used in the quantile estimation. \code{bw} can also be a character string giving a rule to choose the bandwidth. In this case, only option available is \code{\link[=bw.q.SBCcv]{bw.q.SBCcv}}.}

\item{adjust}{A numeric value representing the manual adjustment factor for the bandwidth. Default is 1.}

\item{kernel}{A character vector specifying the kernel function. Available options: \code{"gaussian"}, \code{"epanechnikov"}, \code{"rectangular"}, \code{"triangular"}, \code{"biweight"}, \code{"cosine"} and \code{"optcosine"}.}

\item{from}{Numeric value specifying the lower bound for  evaluation when \code{x} is not provided. Default is 0.01.}

\item{to}{Numeric value specifying the upper bound for  evaluation when \code{x} is not provided. Default is 0.99.}

\item{nb}{An integer specifying the number of points for evaluation when \code{x} is not provided. Default is 99.}

\item{plot}{A logical value indicating whether to plot the quantile estimation. Default is \code{TRUE}.}

\item{...}{Additional arguments to be passed to bandwidth selection functions.}
}
\value{
A list with the following components:
\item{\code{x}}{The points where the quantiles are estimated.}
\item{\code{est_values}}{The estimated quantile values.}
\item{\code{bw}}{The bandwidth used.}
\item{\code{n}}{The sample size after elimination of missing values.}
\item{\code{call}}{The call which produced the result.}
\item{\code{data.name}}{The deparsed name of the y argument (biased dataset).}
\item{\code{has.na}}{Logical; indicates whether the original vector \code{y} contains any \code{NA} values.}
}
\description{
This function estimates the quantile function using the kernel method. It is a generalization of the quantile function estimator given in \insertCite{parzen1979;textual}{WData}  for unbiased data.
}
\details{
Estimator is given by
\deqn{\widehat{F_{h}^{-1}} (\tau) = \int_{0}^{1} \widehat{F^{-1}_n} (t) K_h \left(t-\tau\right) dt,}
where \eqn{\widehat{F^{-1}_n}} is the the empirical quantile function based on the estimator proposed by \insertCite{cox2005;textual}{WData} for the distribution function, \eqn{h} is the bandwidth, \eqn{K} is the kernel density function and \eqn{K_{h}(u)=1/h K\left(u / h\right)}.
}
\examples{
qf.SBCkernel(shrub.data$Width, bw = 0.05, kernel = "epanechnikov")
plot(qf.SBC(y = shrub.data$Width), add = TRUE, col = "red") # For comparison

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=qf.SBC]{qf.SBC}}
}
