% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qf.sen.R
\name{qf.sen}
\alias{qf.sen}
\title{\insertCite{sen1984;textual}{WData} quantile estimator}
\usage{
qf.sen(
  y,
  w = function(y) {
     ifelse(y >= 0, y, NA)
 }
)
}
\arguments{
\item{y}{A numeric vector containing the biased sample.}

\item{w}{A function representing the bias function applied to the data points.
It must be evaluable and positive in each point of the sample \code{y}.
By default, it is set to the length-biased function.}
}
\value{
A function of class \code{eqf}, inheriting from the \code{\link[stats:stepfun]{stepfun}} class, and hence inheriting a \code{\link[stats:stepfun]{knots}} method.
}
\description{
This function computes \insertCite{sen1984;textual}{WData} quantile estimator given a sample and the corresponding biased function.
}
\details{
\insertCite{sen1984;textual}{WData} quantile estimator is expressed as
\deqn{ \widehat{F^{-1}_{\mathrm{S}}}(\tau) = Y_{(i)},
\quad \text{where} \quad i = \max \left\{ i: \sum_{j=1}^{i} \frac{1}{Y_{(j)}} \bigg/ \sum_{k=1}^{n} \frac{1}{Y_{(k)}} \leq \tau  \right\}.}
If the sample size \eqn{n} is small and \eqn{\tau} is very close to zero,
the inequality may not hold for any \eqn{i}, in which case \eqn{i} is taken as 1.
}
\examples{
qf.sen(y = shrub.data$Width)
}
\references{
\insertAllCited{}
}
