% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{quantile_cross_validation}
\alias{quantile_cross_validation}
\title{Quantile cross-validation}
\usage{
quantile_cross_validation(Tis, vals_minus_i, ti, vals, hs, kernel)
}
\arguments{
\item{Tis}{Numeric matrix.}

\item{vals_minus_i}{Numeric matrix.}

\item{ti}{Numeric vector.}

\item{vals}{Numeric vector.}

\item{hs}{Numeric vector of bandwidths.}

\item{kernel}{Character, kernel name.}
}
\value{
Numeric vector with cross-validation scores.
}
\description{
Performs quantile-based cross-validation for bandwidth selection.
}
