% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{shrub.data}
\alias{shrub.data}
\title{Shrub data}
\format{
A \code{data.frame} with the following variables:
\tabular{ll}{
\code{Replica} \tab {Replica identifier (I or II).} \cr \tab \cr
\code{Transect} \tab {Transect identifier (1,2 or 3).} \cr \tab \cr
\code{Number} \tab {Shrub/clump identifier.} \cr \tab \cr
\code{Intercept} \tab {Length of the intersection of the clump of overlapping shrubs with the transect.} \cr \tab \cr
\code{Widht} \tab {Width between two lines tangent to the shrub and parallel to the transect. } \cr \tab \cr
\code{Height} \tab {Maximum height of the shrub encountered by the transect.} \cr \tab \cr
\code{Stems} \tab {Number of stems on the shrub encountered by the transect.} \cr
}
}
\source{
\url{https://www.proquest.com/docview/303721613?\%20T&fromopenview=true&pq-origsite=gscholar&sourcetype=Dissertations\%20}
}
\usage{
data(shrub.data)
}
\description{
Size of the \emph{Cercocarpus montanus} species in an ancient quarry.
}
\details{
During the fall semester of 1986, members of a graduate class on biological sampling techniques taught by Lyman L. McDonald at the University of Wyoming conducted sampling using the linear transect method to study the size of the \emph{Cercocarpus montanus} species in an old limestone quarry located just east of Laramie, Wyoming. In this area, rock fissures run from north to south. Vegetation is more abundant in these fissures, so a baseline was established approximately parallel to them. Transects were laid out perpendicular to this baseline. To ensure uniform coverage over the study area, two independent replicas (I and II) were established, each with 3 parallel transects, positioned to divide the baseline into three equal parts. Consequently, the sampling involved a total of six transects extending from the baseline to the eastern border of the quarry.

Students walked the transects identifying \emph{Cercocarpus montanus} individuals intersected by the transect. For each shrub, its maximum height, width (defined as the maximum distance between two tangents to the shrub's contour parallel to the transect) and the number of main branches were measured. Since the species in question is rhizomatous and adjacent shrubs may be interconnected through root systems, a shrub was defined as an individual if it had a distinctive group of stems at the base and was more than fifteen centimeters from its nearest neighbor. Additionally, measurements were taken of the length of intersection with the transect for shrub clusters. Details of the sampling and measurements for the rest of the characteristics can be found in \insertCite{muttlak1988;textual}{WData}.

Given the sampling method, it is clear that wider shrubs were more likely to be intersected by one of the transects. Consequently, the widths of shrubs collected by the students represent a biased sample with longitudinal bias. Measurements for shrub height and branch number are also subject to bias, but the bias function  is more complicated because it depends on the relationship between the shrub's width and its height and number of branches, respectively.
}
\examples{
data(shrub.data)
names(shrub.data)
str(shrub.data)
class(shrub.data)
}
\references{
\insertAllCited{}
}
\keyword{datasets}
