% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.akbarikernel.R
\name{sp.akbarikernel}
\alias{sp.akbarikernel}
\title{\insertCite{akbari2019;textual}{WData} kernel sparsity estimator}
\usage{
sp.akbarikernel(
  y,
  w = function(y) {
     ifelse(y >= 0, y, NA)
 },
  x,
  bw = NULL,
  adjust = 1,
  kernel = c("gaussian", "epanechnikov", "rectangular", "triangular", "biweight",
    "cosine", "optcosine"),
  from = 0.01,
  to = 0.99,
  nb = 99,
  plot = TRUE
)
}
\arguments{
\item{y}{A numeric vector containing the biased data.}

\item{w}{A function representing the bias function applied to the data points. It must be evaluable and positive in each point of \code{y}. By default, it is set to the length-biased function.}

\item{x}{A numeric vector specifying the points where the sparsity is estimated. Alternatively, \code{from}, \code{to} and \code{nb} can be used to define the evaluation points.}

\item{bw}{The smoothing bandwidth to be used in the sparsity estimation.}

\item{adjust}{A numeric value representing the manual adjustment factor for the bandwidth. Default is 1.}

\item{kernel}{A character vector specifying the kernel function. Available options: \code{"gaussian"}, \code{"epanechnikov"}, \code{"rectangular"}, \code{"triangular"}, \code{"biweight"}, \code{"cosine"} and \code{"optcosine"}.}

\item{from}{Numeric value specifying the lower bound for  evaluation when \code{x} is not provided. Default is 0.01.}

\item{to}{Numeric value specifying the upper bound for  evaluation when \code{x} is not provided. Default is 0.99.}

\item{nb}{An integer specifying the number of points for evaluation when \code{x} is not provided. Default is 99.}

\item{plot}{A logical value indicating whether to plot the sparsity estimation. Default is \code{TRUE}.}
}
\value{
A list with the following components:
\item{\code{x}}{The points where the sparsity is estimated.}
\item{\code{est_values}}{The estimated sparsity values.}
\item{\code{bw}}{The bandwidth used.}
\item{\code{n}}{The sample size after elimination of missing values.}
\item{\code{call}}{The call which produced the result.}
\item{\code{data.name}}{The deparsed name of the y argument (biased dataset).}
\item{\code{has.na}}{Logical; indicates whether the original vector \code{y} contains any \code{NA} values.}
}
\description{
This function estimates the sparsity of a distribution using \insertCite{akbari2019;textual}{WData} kernel estimator.
}
\details{
\insertCite{akbari2019;textual}{WData} sparsity estimator is given by
\deqn{\widehat{\left(F^{-1}\right)_{h}^{(1)}} (\tau) = \int_{0}^{1} K_h (t-\tau) d \widehat{F^{-1}_{\operatorname{SEN}}} (t),}
where \eqn{\widehat{F^{-1}_{\operatorname{SEN}}}} is \insertCite{sen1984;textual}{WData} quantile estimator, \eqn{h} is the bandwidth, \eqn{K} is the kernel density function and \eqn{K_{h}(u)=1/h K\left(u / h\right)}.
}
\examples{
sp.akbarikernel(shrub.data$Width, bw = 0.05, kernel = "epanechnikov")

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=qf.sen]{qf.sen}}
}
