% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.akbaripi.R
\name{sp.akbaripi}
\alias{sp.akbaripi}
\title{\insertCite{akbari2019;textual}{WData} plug-in sparsity estimator}
\usage{
sp.akbaripi(
  y,
  w = function(y) {
     ifelse(y >= 0, y, NA)
 },
  bw = "bw.f.BGMnrd0",
  adjust = 1,
  kernel = c("gaussian", "epanechnikov", "rectangular", "triangular", "biweight",
    "cosine", "optcosine")
)
}
\arguments{
\item{y}{A numeric vector containing the biased data.}

\item{w}{A function representing the bias function applied to the data points. It must be evaluable and positive in each point of \code{y}. By default, it is set to the length-biased function.}

\item{bw}{The smoothing bandwidth to be used in the density estimation. \code{bw} can also be a character string giving a rule to choose the bandwidth. In this case, options available are \code{\link[=bw.f.BGMnrd0]{bw.f.BGMnrd0}}, \code{\link[=bw.f.BGMcv]{bw.f.BGMcv}},  \code{\link[=bw.f.BGMboot1]{bw.f.BGMboot1}} and \code{\link[=bw.f.BGMboot2]{bw.f.BGMboot2}}.
Default is \code{\link[=bw.f.BGMnrd0]{bw.f.BGMnrd0}}. The specified (or computed) value of \code{bw} is multiplied by \code{adjust}.}

\item{adjust}{A numeric value representing the manual adjustment factor for the bandwidth. Default is 1.}

\item{kernel}{A character vector specifying the kernel function. Available options: \code{"gaussian"}, \code{"epanechnikov"}, \code{"rectangular"}, \code{"triangular"}, \code{"biweight"}, \code{"cosine"} and \code{"optcosine"}.}
}
\value{
A function of class \code{eqf}, inheriting from the \code{\link[stats:stepfun]{stepfun}} class, and hence inheriting a \code{\link[stats:stepfun]{knots}} method.
}
\description{
This function estimates the sparsity of a distribution using \insertCite{akbari2019;textual}{WData} plug-in sparsity estimator.
}
\details{
\insertCite{akbari2019;textual}{WData} suggests estimating the sparsity function using the following plug-in estimator:
\deqn{\widehat{\left(F^{-1}\right)_{\operatorname{PI}, h}^{(1)}} (\tau) =
\frac{1}{\widehat{f}_{\operatorname{J}, h}\left( \widehat{F^{-1}_{\operatorname{SEN}}} (\tau )\right)},}
where \eqn{\widehat{f}_{\operatorname{J}, h}} is \insertCite{jones1991;textual}{WData} kernel density estimator and \eqn{\widehat{F^{-1}_{\operatorname{SEN}}}} is \insertCite{sen1984;textual}{WData} quantile function estimator.
}
\examples{
sp.akbaripi(shrub.data$Width, bw = 0.25)
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=qf.sen]{qf.sen}}, \code{\link[=df.jones]{df.jones}}
}
