% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{spnorthdatras.data}
\alias{spnorthdatras.data}
\title{\emph{Conger conger} 2024 ICES \emph{Spanish North Coast Bottom Trawl Survey} (DATRAS)}
\format{
A \code{data.frame} with the following variables:
\tabular{ll}{
\code{StNo} \tab National station code or number.\cr
\code{HaulNo} \tab Haul number.\cr
\code{TotalNo} \tab Total number of individuals in the catch.\cr
\code{SubFactor} \tab Subsampling factor (\code{SubWgt / CatCatchWgt}).\cr
\code{SubWgt} \tab Subsample weight (grams).\cr
\code{CatCatchWgt} \tab Catch weight per category (grams).\cr
\code{LngtClass} \tab Length (centimetres).\cr
\code{HLNoAtLngt} \tab Number of individuals at length \code{LngtClass}.\cr
}
}
\source{
\href{https://datras.ices.dk/Data_products/Download/Download_Data_public.aspx}{ICES DATRAS Download web portal}

\href{https://ices-library.figshare.com/articles/report/SISP_15_-_Manual_of_the_IBTS_North_Eastern_Atlantic_Surveys/19051037}{Manual of the IBTS North Eastern Atlantic Surveys}

\href{https://vocab.ices.dk}{ICES DATRAS Field Descriptions}

\href{https://www.marinespecies.org}{World Register of Marine Species (WoRMS)}

\href{https://cran.r-project.org/web/packages/icesVocab/index.html}{icesVocab R package: ICES Vocabularies Database Web Services}
}
\usage{
data(spnorthdatras.data)
}
\description{
\emph{Conger conger} 2024 dataset from the ICES \emph{Spanish North Coast Bottom Trawl Survey}, extracted from DATRAS (Database of Trawl Surveys).
}
\details{
DATRAS is maintained by ICES (International Council for the Exploration of the Sea) to collate, document, and standardise trawl survey data, ensure data quality, and facilitate access for stock assessment and fish community studies. It primarily contains data from bottom trawl surveys coordinated by ICES expert groups, covering the Baltic Sea, Skagerrak, Kattegat, North Sea, English Channel, Celtic Sea, Irish Sea, Bay of Biscay, and the eastern Atlantic from the Shetlands to Gibraltar. A complete list of ICES DATRAS surveys is available \href{https://datras.ices.dk/home/surveycode.aspx}{here}.

\code{spnorthdatras.data} includes \emph{Conger conger} 2024 length data from the ICES \emph{Spanish North Coast Bottom Trawl Survey} (DATRAS). Data can be manually downloaded from the \insertCite{ices_datras;textual}{WData} \href{https://datras.ices.dk/Data_products/Download/Download_Data_public.aspx}{web portal} or  obtained using the \code{icesDatras} R package as follows:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{icesDatras::getDATRAS(
 record = "HL",
 survey = "SP-NORTH",
 year = 2024,
 quarter = 1:4
) \%>\%
dplyr::filter(Valid_Aphia == 126285)
}\if{html}{\out{</div>}}

Fishing gears are selective by species and, within species, by size. The probability that a fish is retained and hauled to the surface depends on its body size because of mesh selectivity. Although retention is more directly determined by girth, length is typically recorded in practice, and the two measures are strongly linearly related. As a result, the observed total length distributions are biased toward larger individuals because of size-dependent trawl selectivity (\insertCite{wileman1996;textual}{WData}).

Missing values in the original DATRAS files (\code{-9}) were replaced by  \code{NA}. Some variables were converted to factors, and several fields from the original database were omitted; otherwise, the column names and formats remained aligned with the original specification.
}
\examples{
data(spnorthdatras.data)
names(spnorthdatras.data)
str(spnorthdatras.data)
class(spnorthdatras.data)
}
\references{
\insertAllCited{}
}
\keyword{datasets}
