% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_boxplot.R
\name{w_boxplot}
\alias{w_boxplot}
\alias{w_boxplot.default}
\alias{w_boxplot.formula}
\title{Box plot}
\usage{
w_boxplot(x, ...)

\method{w_boxplot}{default}(
  x,
  w = function(y) {
     ifelse(y >= 0, y, NA)
 },
  qmethod = "qf.SBC",
  ...,
  range = 1.5,
  width = NULL,
  varwidth = FALSE,
  notch = FALSE,
  outline = TRUE,
  names,
  plot = TRUE,
  border = par("fg"),
  col = "lightgray",
  log = "",
  pars = list(boxwex = 0.8, staplewex = 0.5, outwex = 0.5),
  ann = !add,
  horizontal = FALSE,
  add = FALSE,
  at = NULL
)

\method{w_boxplot}{formula}(
  formula,
  data = NULL,
  w = function(y) {
     ifelse(y >= 0, y, NA)
 },
  qmethod = c("qf.SBC", "qf.sen"),
  ...,
  subset,
  na.action = NULL,
  xlab = mklab(y_var = horizontal),
  ylab = mklab(y_var = !horizontal),
  add = FALSE,
  ann = !add,
  horizontal = FALSE,
  drop = FALSE,
  sep = ".",
  lex.order = FALSE
)
}
\arguments{
\item{x}{For specifying data from which the boxplots are to be produced. Either a numeric vector, or a single list containing such vectors. Additional unnamed arguments specify further data as separate vectors (each corresponding to a component boxplot). \code{NA}s are allowed in the data. Additional unnamed arguments specify further data as separate vectors.}

\item{...}{For the formula method, named arguments to be passed to the default method. For the default method, unnamed arguments are additional data vectors (unless \code{x} is a list when they are ignored), and named arguments are arguments and graphical parameters to be passed to \code{bxp} in addition to the ones given by argument pars (and override those in pars). Note that \code{bxp} may or may not make use of graphical parameters it is passed: see its documentation.}

\item{w}{A function representing the bias function applied to the data points. It must be evaluable and positive in each point of \code{x}. By default, it is set to the length-biased function.}

\item{qmethod}{Character string specifying the quantile estimator to be used. Options are \code{"qf.SBC"} (\code{\link[=qf.SBC]{qf.SBC}}) and \code{"qf.sen"} (\code{\link[=qf.sen]{qf.sen}}). Default is \code{"qf.SBC"}.}

\item{range}{This determines how far the plot whiskers extend out from the box. If range is positive, the whiskers extend to the most extreme data point which is no more than range times the interquartile range from the box. A value of zero causes the whiskers to extend to the data extremes.}

\item{width}{A vector giving the relative widths of the boxes making up the plot.}

\item{varwidth}{If \code{varwidth} is \code{TRUE}, the boxes are drawn with widths proportional to the square-roots of the number of observations in the groups.}

\item{notch}{If notch is \code{TRUE}, a notch is drawn in each side of the boxes.}

\item{outline}{If outline is not true, the outliers are not drawn (as points whereas S+ uses lines).}

\item{names}{Character vector or expressions for group labels.}

\item{plot}{If \code{TRUE} (the default) then a boxplot is produced. If not, the summaries which the boxplots are based on are returned.}

\item{border}{An optional vector of colors for the outlines of the boxplots. The values in border are recycled if the length of border is less than the number of plots.}

\item{col}{If col is non-null it is assumed to contain colors to be used to color the bodies of the box plots. By default they are in the background color.}

\item{log}{Character indicating if x or y or both coordinates should be plotted in log scale.}

\item{pars}{A list of (potentially many) more graphical parameters, e.g., \code{boxwex} or \code{outpch}; these are passed to \code{bxp} (if plot is \code{TRUE}); for details, see \code{\link[graphics:par]{par}}.}

\item{ann}{Logical; if \code{TRUE}, axes are annotated (by \code{xlab} and \code{ylab}).}

\item{horizontal}{Logical indicating if the boxplots should be horizontal; default \code{FALSE} means vertical boxes.}

\item{add}{Logical, if \code{TRUE} add boxplot to current plot.}

\item{at}{Numeric vector giving the locations where the boxplots should be drawn, particularly when \code{add = TRUE}; defaults to \code{1:n} where \code{n} is the number of boxes.}

\item{formula}{A formula, such as \code{y ~ grp}, where \code{y} is a numeric vector of data values to be split into groups according to the grouping variable \code{grp} (usually a factor). Note that \code{~ g1 + g2} is equivalent to \code{g1:g2}.}

\item{data}{A \code{data.frame} (or list) from which the variables in \code{formula} should be taken.}

\item{subset}{An optional vector specifying a subset of observations to be used for plotting.}

\item{na.action}{A function indicating what should happen when the data contain \code{NA}s. The default is to ignore missing values in either the response or the group.}

\item{xlab, ylab}{X- and y-axis annotation. Can be suppressed by \code{ann = FALSE}.}

\item{drop, sep, lex.order}{Passed to \code{\link[base:split]{split.default}}.}

\item{boxwex}{A scale factor to be applied to all boxes. When there are only a few groups, the appearance of the plot can be improved by making the boxes narrower.}

\item{staplewex}{Staple line width expansion, proportional to box width.}

\item{outwex}{Outlier line width expansion, proportional to box width.}
}
\value{
A list with components:
\item{stats}{A matrix, each column contains the extreme of the lower whisker, the lower hinge, the median, the upper hinge and the extreme of the upper whisker for one group/plot. If all the inputs have the same class attribute, so will this component.}
\item{n}{A vector with the number of (non-\code{NA}) observations in each group.}
\item{conf}{A matrix where each column contains the lower and upper extremes of the notch.}
\item{out}{The values of any data points which lie beyond the extremes of the whiskers.}
\item{group}{A vector of the same length as out whose elements indicate to which group the outlier belongs.}
\item{names}{A vector of names for the groups.}
}
\description{
This function produces a box-and-whisker plot(s) of the given (grouped) values for biased data.
}
\details{
This function is a modification of the \code{\link[graphics:boxplot]{boxplot}} function, which is used to create boxplots for biased data by calculating the five-number summary of the data based on \code{\link[=qf.sen]{qf.sen}} quantile estimator or in \code{\link[=qf.SBC]{qf.SBC}} quantile estimator.
The generic function \code{\link[=w_boxplot]{w_boxplot}} has a default method (\code{\link[=w_boxplot.default]{w_boxplot.default}}) and a formula
interface (\code{\link[=w_boxplot.formula]{w_boxplot.formula}}). Parallel boxplots are plotted for multiple groups.
Missing values are ignored when forming boxplots.
}
\examples{
w_boxplot(Width ~ Replica, data = shrub.data, qmethod = "qf.SBC", main = "Boxplot by Group")
w_boxplot(shrub.data$Width, qmethod = "qf.SBC", horizontal = TRUE, main = "Boxplot")
w_boxplot(Width ~ Replica, data = shrub.data, qmethod = "qf.sen", main = "Boxplot by Group")
w_boxplot(shrub.data$Width, qmethod = "qf.sen", horizontal = TRUE, main = "Boxplot")
}
\seealso{
\code{\link[=w_boxstats]{w_boxstats}}
}
