% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_boxplot.stats.R
\name{w_boxstats}
\alias{w_boxstats}
\title{Box plot statistics}
\usage{
w_boxstats(
  x,
  w = function(y) {
     ifelse(y >= 0, y, NA)
 },
  qmethod = c("qf.SBC", "qf.sen"),
  coef = 1.5,
  do.conf = TRUE,
  do.out = TRUE
)
}
\arguments{
\item{x}{A numeric vector for which the boxplot will be constructed (\code{NA}s and \code{NaN}s are allowed and omitted).}

\item{w}{A function representing the bias function applied to the data points.
It must be evaluable and positive in each point of \code{y}.
By default, it is set to the length-biased function.}

\item{qmethod}{Character string specifying the quantile estimator to be used. Options are \code{"qf.SBC"} (\code{\link[=qf.SBC]{qf.SBC}}) and \code{"qf.sen"} (\code{\link[=qf.sen]{qf.sen}}). Default is \code{"qf.SBC"}.}

\item{coef}{This determines how far the plot ‘whiskers’ extend out from the box. If \code{coef} is positive, the whiskers extend to the most extreme data point which is no more than \code{coef} times the length of the box away from the box. A value of zero causes the whiskers to extend to the data extremes (and no outliers be returned).}

\item{do.conf}{Logical; if \code{FALSE}, it will be empty in the result.}

\item{do.out}{Logical; if \code{FALSE}, it will be empty in the result.}
}
\value{
A list with components:
\item{stats}{A vector of length 5, containing the extreme of the lower whisker, the lower ‘hinge’, the median, the upper ‘hinge’ and the extreme of the upper whisker.}
\item{n}{The number of non-\code{NA} observations in the sample.}
\item{conf}{The lower and upper extremes of the ‘notch’ (\verb{if(do.conf)}). See the details on \code{\link[=boxplot.stats]{boxplot.stats}}.}
\item{out}{The values of any data points which lie beyond the extremes of the whiskers (\verb{if(do.out)}).}
}
\description{
This function is typically called by another function to gather the statistics necessary for producing box plots, but may be invoked separately.
}
\details{
This function is a modification of the \code{\link[=boxplot.stats]{boxplot.stats}} function, which is used to calculate the statistics needed to create a boxplot. The main difference is that this function allows to compute a boxplot for biased data by calculating the five-number summary of the data based on \code{\link[=qf.sen]{qf.sen}} quantile estimator or in \code{\link[=qf.SBC]{qf.SBC}} quantile estimator.
}
\examples{
w_boxstats(shrub.data$Width)
}
\seealso{
\code{\link[=w_fivenum]{w_fivenum}}, \code{\link[=w_boxplot]{w_boxplot}}.
}
