% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_fivenum.R
\name{w_fivenum}
\alias{w_fivenum}
\title{Tukey five-number summary}
\usage{
w_fivenum(
  x,
  w = function(y) {
     ifelse(y >= 0, y, NA)
 },
  qmethod = c("qf.SBC", "qf.sen"),
  na.rm = TRUE
)
}
\arguments{
\item{x}{Numeric, maybe including \code{NA}s and \code{Inf}s.}

\item{w}{A function representing the bias function applied to the data points.
It must be evaluable and positive in each point of \code{x}. By default, it is set to the length-biased function.}

\item{qmethod}{Character string specifying the quantile estimator to be used. Options are \code{"qf.SBC"} (\code{\link[=qf.SBC]{qf.SBC}}) and \code{"qf.sen"} (\code{\link[=qf.sen]{qf.sen}}). Default is \code{"qf.SBC"}.}

\item{na.rm}{Logical; if \code{TRUE}, all \code{NA}s and \code{NaN}s are dropped, before the statistics are computed.}
}
\value{
A numeric vector of length 5 containing the Tukey five number summary (minimum, lower-hinge, median, upper-hinge, maximum) for biased data.
}
\description{
This function calculates Tukey's five number summary (minimum, lower-hinge, median, upper-hinge, maximum) for biased data.
}
\details{
This function is a modification of the \code{\link[stats:fivenum]{fivenum}} function, which is used to calculate Tukey's five number summary. The main difference is that this function allows to compute the five-number summary for biased data by calculating the five-number summary of the data based on \code{\link[=qf.sen]{qf.sen}} quantile estimator or in \code{\link[=qf.SBC]{qf.SBC}} quantile estimator.
}
\examples{
w_fivenum(shrub.data$Width)
}
\seealso{
\code{\link[=w_boxstats]{w_boxstats}}.
}
