% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_qqnorm.R
\name{w_qqnorm}
\alias{w_qqnorm}
\alias{w_qqnorm.default}
\alias{w_qqline}
\title{Quantile-quantile plots}
\usage{
w_qqnorm(y, ...)

\method{w_qqnorm}{default}(
  y,
  w = function(y) {
     ifelse(y >= 0, y, NA)
 },
  ylim,
  main = "Normal Q-Q Plot",
  xlab = "Theoretical Quantiles",
  ylab = "Sample Quantiles",
  plot.it = TRUE,
  datax = FALSE,
  ...
)

w_qqline(
  y,
  w = function(y) {
     ifelse(y >= 0, y, NA)
 },
  datax = FALSE,
  distribution = qnorm,
  probs = c(0.25, 0.75),
  qmethod = c("qf.SBC", "qf.sen"),
  ...
)
}
\arguments{
\item{y}{The biased data.}

\item{...}{Additional graphical parameters.}

\item{w}{A function representing the bias function applied to the data points. It must be evaluable and positive in each point of \code{y}. By default, it is set to the length-biased function.}

\item{ylim}{The limits for the axis.}

\item{main}{An overall title for the plot.}

\item{xlab}{A label for the x axis.}

\item{ylab}{A label for the y axis.}

\item{plot.it}{A logical value indicating whether to plot the Q-Q plot. Default is \code{TRUE}.}

\item{datax}{A logical value indicating whether the data should be plotted on the x-axis. Default is \code{FALSE}.}

\item{distribution}{Quantile function for reference theoretical distribution.}

\item{probs}{Numeric vector of length two, representing probabilities. Corresponding quantile pairs define the line drawn.}

\item{qmethod}{Character string specifying the quantile estimator to be used. Options are \code{"qf.SBC"} (\code{\link[=qf.SBC]{qf.SBC}}) and \code{"qf.sen"} (\code{\link[=qf.sen]{qf.sen}}). Default is \code{"qf.SBC"}.}
}
\value{
For \code{w_qqnorm}, a list with components:
\item{x}{The \code{x} coordinates of the points that were/would be plotted}
\item{y}{The original \code{y} vector, i.e., the corresponding \code{y} coordinates including \code{NA}s.}
}
\description{
This function produces a normal Q-Q plot of the values in \code{y} with the bias function \code{w}.
}
\details{
This function is a modification of the \code{\link[stats:qqnorm]{qqnorm}} function, which is used to create boxplots for biased data. \code{\link[=w_qqnorm]{w_qqnorm}} is a generic function whose default method produces a normal Q-Q plot of the values in \code{y}. \code{\link[=w_qqline]{w_qqline}} adds a line to a theoretical Q-Q plot, by default normal, which passes through the quantiles defined in \code{probs}, by default the first and third quartiles.
}
\examples{
w_qqnorm(shrub.data$Width)
w_qqline(shrub.data$Width, col = "blue", qmethod = "qf.sen")
w_qqline(shrub.data$Width, col = "magenta", qmethod = "qf.SBC")
}
\seealso{
\code{\link[=qf.SBC]{qf.SBC}}, \code{\link[=qf.sen]{qf.sen}}
}
