% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw.F.BD.R
\name{bw.F.BD}
\alias{bw.F.BD}
\title{\insertCite{bose2022;textual}{WData} bandwidth selection for \insertCite{bose2022;textual}{WData} kernel distribution estimator}
\usage{
bw.F.BD(
  y,
  w = function(y) {
     ifelse(y >= 0, y, NA)
 },
  x,
  c_adj
)
}
\arguments{
\item{y}{A numeric vector containing the biased data.}

\item{w}{A function representing the bias function to be used in the estimation. It must be evaluable and positive in each point of \code{y}. By default, it is set to the length-biased function.}

\item{x}{A numeric vector containing the points for which the bandwidth is calculated.}

\item{c_adj}{A numeric vector representing the constants to be used in the bandwidth calculation for each point of \code{x}.}
}
\value{
A numeric vector containing the bandwidths for each point in \code{x}.
}
\description{
This function calculates the bandwidth for \insertCite{bose2022;textual}{WData} distribution estimator following the method proposed by \insertCite{bose2022;textual}{WData}.
}
\details{
Bandwidths are calculated using the formula:
\deqn{\widehat{h}_{F, \mathrm{BD}, C(y)} (y) = \frac{C(y) \widehat{\sigma}_{w}}{(n w(y))^{1/3}},}
where \eqn{C(y)} is a positive parameter that depends on the point \eqn{y} and \eqn{\widehat{\sigma}_{w}} is an estimate of the standard deviation of the distribution given by
\deqn{\widehat{\sigma}_w=\sqrt{\left(\frac{1}{n} \sum_{i=1}^n \frac{1}{w(Y_i)}\right)^{-1}
\left[\left(\frac{1}{n} \sum_{i=1}^n w(Y_i)\right)-\left(\frac{1}{n} \sum_{i=1}^n \frac{1}{w(Y_i)}\right)^{-1}\right]}.}

The parameter \eqn{C(y)} is provided to the function by using the argument \code{c_adj}, which is a vector of positive values that is used to compute the bandwidth for each point in \code{x}.
The simulations carried out by \insertCite{bose2022;textual}{WData} suggest that choosing \eqn{C(y)=0.25} or \eqn{C(y)=0.5} provides good results in the tail region of the distribution, with tails defined as points below the 5th percentile or above the 95th percentile. On the other hand, \eqn{C(y)=1.3} provides good results for the remaining points.
If some bandwidths are not positive, they are replaced by the mean of the neighbors.
}
\examples{
bw.F.BD(shrub.data$Width, x = seq(0, 1, length.out = 512), c_adj = rep(1, 512))
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=cdf.bd]{cdf.bd}}
}
