% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw.F.BD.R
\name{bw.F.BD}
\alias{bw.F.BD}
\title{\insertCite{bose2022;textual}{WData} bandwidth selection for \insertCite{bose2022;textual}{WData} kernel distribution estimator}
\usage{
bw.F.BD(
  y,
  w = function(y) {
     ifelse(y >= 0, y, NA)
 },
  x,
  c_adj
)
}
\arguments{
\item{y}{A numeric vector containing the biased data.}

\item{w}{A function representing the bias function to be used in the estimation. It must be evaluable and positive in each point of \code{y}. By default, it is set to the length-biased function.}

\item{x}{A numeric vector containing the points for which the bandwidth is calculated.}

\item{c_adj}{A numeric vector representing the constants to be used in the bandwidth calculation for each point of \code{x}.}
}
\value{
A numeric vector containing the bandwidths for each point in \code{x}.
}
\description{
This function calculates the bandwidth for \insertCite{bose2022;textual}{WData} distribution estimator following the method proposed by \insertCite{bose2022;textual}{WData}.
}
\details{
\insertCite{bose2022;textual}{WData} propose using local bandwidths the form
\deqn{\widehat{h}_{\mathrm{BD}, C(y)} (y) = \frac{C(y) \widehat{\sigma}}{(n w(y))^{1/3}},}
where \eqn{C(y)} is a positive function that depends on the point \eqn{y}, and \eqn{\widehat{\sigma}} is an estimate of the standard deviation of the distribution. \insertCite{borrajo2017;textual}{WData} suggests two options for estimating the standard deviation \eqn{\sigma}. The first is to take the square root of the following estimator of \eqn{\sigma^2}:
\deqn{\widehat{\sigma}_w^2=\left(\frac{1}{n} \sum_{i=1}^n \frac{1}{w(Y_i)}\right)^{-1}\left[\left(\frac{1}{n} \sum_{i=1}^n w(Y_i)\right)-\left(\frac{1}{n} \sum_{i=1}^n \frac{1}{w(Y_i
)}\right)^{-1}\right].}
The second option is to consider a robust estimator of \eqn{\sigma}, such as
\deqn{\widehat{\sigma}_{\mathrm{IQR}}=\mathrm{IQR}/(\Phi(0.75)-\Phi(0.25)),}
where \eqn{\mathrm{IQR}} is the interquartile range of the sample \eqn{\{Y_1, Y_2, \dots, Y_n\}} and \eqn{\Phi} is the cumulative distribution function of the standard normal distribution.
In order to prevent oversmoothing in the estimation, this
function considers \eqn{\widehat{\sigma}= \min\{\widehat{\sigma}_w, \widehat{\sigma}_{\mathrm{IQR}}\}}.
If some bandwidths are not positive, they are replaced by the mean of the neighbors.

The simulations carried out by \insertCite{bose2022;textual}{WData} suggest that choosing \eqn{C(y)=0.25} or \eqn{C(y)=0.5} provides good results in the tail region of the distribution, with tails defined as points below the 5th percentile or above the 95th percentile. On the other hand, \eqn{C(y)=1.3} provides good results for the remaining points.
}
\examples{
bw.F.BD(shrub.data$Width, x = seq(0, 1, length.out = 512), c_adj = rep(1, 512))
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=cdf.bd]{cdf.bd}}
}
