% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.jones.R
\name{df.jones}
\alias{df.jones}
\title{\insertCite{jones1991;textual}{WData} kernel density estimator}
\usage{
df.jones(
  y,
  w = function(y) {
     ifelse(y >= 0, y, NA)
 },
  x,
  bw = "bw.f.BGMnrd0",
  adjust = 1,
  kernel = c("gaussian", "epanechnikov", "rectangular", "triangular", "biweight",
    "cosine", "optcosine"),
  from,
  to,
  nb = 512L,
  plot = TRUE,
  ...
)
}
\arguments{
\item{y}{A numeric vector containing the biased data.}

\item{w}{A function representing the bias function applied to the data points.
It must be evaluable and positive in each point of \code{y}.
By default, it is set to the length-biased function.}

\item{x}{A numeric vector specifying the points where the density is estimated. Alternatively, \code{from}, \code{to} and \code{nb} can be used to define the evaluation points.}

\item{bw}{The smoothing bandwidth to be used in the density estimation. \code{bw} can also be a character string giving a rule to choose the bandwidth. In this case, options available are \code{\link[=bw.f.BGMnrd0]{bw.f.BGMnrd0}}, \code{\link[=bw.f.BGMcv]{bw.f.BGMcv}},  \code{\link[=bw.f.BGMboot1]{bw.f.BGMboot1}} and \code{\link[=bw.f.BGMboot2]{bw.f.BGMboot2}}.
Default is \code{\link[=bw.f.BGMnrd0]{bw.f.BGMnrd0}}. The specified (or computed) value of \code{bw} is multiplied by \code{adjust}.}

\item{adjust}{A numeric value representing the manual adjustment factor for the bandwidth. Default is 1.}

\item{kernel}{A character vector specifying the kernel function. Available options: \code{"gaussian"}, \code{"epanechnikov"}, \code{"rectangular"}, \code{"triangular"}, \code{"biweight"}, \code{"cosine"} and \code{"optcosine"}.}

\item{from}{Numeric value specifying the lower bound for  evaluation when \code{x} is not provided. Default is calculated based on the range of input data.}

\item{to}{Numeric value specifying the upper bound for  evaluation when \code{x} is not provided. Default is calculated based on the range of input data.}

\item{nb}{An integer specifying the number of points for evaluation when \code{x} is not provided. Default is 512.}

\item{plot}{A logical value indicating whether to plot the density estimation. Default is \code{TRUE}.}

\item{...}{Additional arguments to be passed to bandwidth selection functions.}
}
\value{
A list with the following components:
\item{\code{x}}{The points where the density is estimated.}
\item{\code{est_values}}{The estimated density values.}
\item{\code{bw}}{The bandwidth used.}
\item{\code{n}}{The sample size after elimination of missing values.}
\item{\code{call}}{The call which produced the result.}
\item{\code{data.name}}{The deparsed name of the \code{y} argument (biased dataset).}
\item{\code{has.na}}{Logical; indicates whether the original vector \code{y} contains any \code{NA} values.}
}
\description{
This function calculates \insertCite{jones1991;textual}{WData} kernel density estimator given a biased dataset and its bias function.
}
\details{
\insertCite{jones1991;textual}{WData} kernel density estimator is expressed as
\deqn{\widehat{f}_{\mathrm{J}, h_{f}}(y)=\frac{\widehat{\mu}_w}{n}\sum_{i=1}^{n}  \frac{1}{w(Y_i)} K_{h_{f}}(y-Y_i),\quad
\text{where}\quad \widehat{\mu}_w=n \left(\sum_{i=1}^{n} \frac{1}{w(Y_i)}\right)^{-1},}
\eqn{h_{f}} is the bandwidth, \eqn{K} is the kernel density function and \eqn{K_{h_{f}}(u)=1/h_{f} K\left(u / h_{f}\right)}.
}
\examples{
# Rule of the thumb
df.jones(y = shrub.data$Width, kernel = "epanechnikov", bw = "bw.f.BGMnrd0")
# Cross Validation
df.jones(y = shrub.data$Width, kernel = "epanechnikov", bw = "bw.f.BGMcv")
# Bootstrap
df.jones(y = shrub.data$Width, kernel = "epanechnikov", bw = "bw.f.BGMboot1", bw0 = "RT")
df.jones(y = shrub.data$Width, kernel = "epanechnikov", bw = "bw.f.BGMboot1", bw0 = "Opt")
df.jones(y = shrub.data$Width, kernel = "epanechnikov", bw = "bw.f.BGMboot2", nh = 50L)
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=bw.f.BGMnrd0]{bw.f.BGMnrd0}}, \code{\link[=bw.f.BGMcv]{bw.f.BGMcv}},  \code{\link[=bw.f.BGMboot1]{bw.f.BGMboot1}} , \code{\link[=bw.f.BGMboot2]{bw.f.BGMboot2}}
}
