% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qf.SBC.R
\name{qf.SBC}
\alias{qf.SBC}
\title{Empirical quantile function based on \insertCite{cox2005;textual}{WData} distribution estimator}
\usage{
qf.SBC(
  y,
  w = function(y) {
     ifelse(y >= 0, y, NA)
 }
)
}
\arguments{
\item{y}{A numeric vector containing the biased data.}

\item{w}{A function representing the bias function applied to the data points. It must be evaluable and positive in each point of \code{y}. By default, it is set to the length-biased function.}
}
\value{
A function of class \code{eqf}, inheriting from the \code{\link[stats:stepfun]{stepfun}} class, and hence inheriting a \code{\link[stats:stepfun]{knots}} method.
}
\description{
This function implements the empirical quantile function based on the estimator
proposed by \insertCite{cox2005;textual}{WData} for the distribution function.
}
\details{
The estimator is defined as:
\deqn{ \widehat{F^{-1}_n}(\tau) = Y_{(i)}
\quad \text{where} \quad
i = \min \left\{ i: \tau \leq  \sum_{j=1}^{i} \frac{1}{Y_{(j)}}
\bigg/ \sum_{i=1}^{n} \frac{1}{Y_{(j)}} \right\}.}
The estimator can also be written as:
\deqn{ \widehat{F^{-1}_n} (\tau) = \sum_{i=1}^{n} Y_{(i)} \mathbb{I} \left( T_{i-1} < \tau \leq T_{i} \right),
\quad \text{where}
\quad T_i = \sum_{j=1}^{i} \frac{1}{Y_{(j)}} \bigg/ \sum_{i=1}^{n} \frac{1}{Y_{(j)}}.}
It is understood that \eqn{T_{0} = 0}.
}
\examples{
qf.SBC(y = shrub.data$Width)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=cdf.cox]{cdf.cox}}
}
