% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qf.sen.R
\name{qf.sen}
\alias{qf.sen}
\title{\insertCite{sen1984;textual}{WData} quantile estimator}
\usage{
qf.sen(
  y,
  w = function(y) {
     ifelse(y >= 0, y, NA)
 }
)
}
\arguments{
\item{y}{A numeric vector containing the biased data.}

\item{w}{A function representing the bias function applied to the data points. It must be evaluable and positive in each point of \code{y}. By default, it is set to the length-biased function.}
}
\value{
A function of class \code{eqf}, inheriting from the \code{\link[stats:stepfun]{stepfun}} class, and hence inheriting a \code{\link[stats:stepfun]{knots}} method.
}
\description{
This function calculates \insertCite{sen1984;textual}{WData} quantile estimator.
}
\details{
The estimator proposed by \insertCite{sen1984;textual}{WData} is defined as:
\deqn{ \widehat{F^{-1}_{\mathrm{SEN}}}(\tau) = Y_{(i)},
\quad \text{where} \quad i = \max \left\{ i: \sum_{j=1}^{i} \frac{1}{Y_{(j)}} \bigg/ \sum_{i=j}^{n} \frac{1}{Y_{(j)}} \leq \tau  \right\}.}
If the sample size \eqn{n} is small and \eqn{\tau} is very close to zero,
the inequality may not hold for any \eqn{i}, in which case \eqn{i} is taken as 1.
The estimator can also be written as:
\deqn{ \widehat{F^{-1}_{\mathrm{SEN}}} (\tau) = \sum_{i=1}^{n} Y_{(i)} \mathbb{I} \left( T_i \leq \tau < T_{i+1} \right),
\quad \text{where} \quad
T_i = \sum_{j=1}^{i} \frac{1}{Y_{(j)}} \bigg/ \sum_{i=j}^{n} \frac{1}{Y_{(j)}}.}
It is understood that \eqn{T_{n+1} = 1}.
}
\examples{
qf.sen(y = shrub.data$Width)
}
\references{
\insertAllCited{}
}
