% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WData-package.R
\docType{package}
\name{WData-package}
\alias{WData}
\alias{WData-package}
\title{WData: Statistical Inference for Weighted Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A set of tools for analyzing and modeling data subject to sampling biases. Provides functions to estimate the density and cumulative distribution functions from biased samples of continuous distributions. Includes the estimators proposed by Bhattacharyya et al. (1988) and Jones (1991) for density, and by Cox (2005) and Bose and Dutta (2022) for distribution, with different bandwidth selectors. Also includes a real length-biased dataset on shrub width from Muttlak (1988).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/noeliasanchmrt/WData}
  \item Report bugs at \url{https://github.com/noeliasanchmrt/WData/issues}
}

}
\author{
\strong{Maintainer}: Sánchez Martínez Noelia \email{noelia.sanchez.martinez@rai.usc.es} (\href{https://orcid.org/0000-0002-2869-1347}{ORCID})

Authors:
\itemize{
  \item Borrajo Garcia Maria Isabel \email{mariaisabel.borrajo@usc.es} (\href{https://orcid.org/0000-0002-3372-1993}{ORCID})
  \item Conde Amboage Mercedes \email{mercedes.amboage@usc.es} (\href{https://orcid.org/0000-0003-0306-8142}{ORCID})
}

}
\keyword{internal}
