% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WData-package.R
\docType{package}
\name{WData-package}
\alias{WData}
\alias{WData-package}
\title{WData: Density, Distribution, and Quantile Estimation for Biased Datasets}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Set of tools for analyzing and modeling data that may be subject to biases in sampling. It offers functions to estimate density function and cumulative distribution function from a biased sample of a continuous distribution. Regarding density function estimation, the package includes @bhattacharyya1988 and @jones1991 density estimators and various bandwidth selectors for the latter, enhancing the flexibility and adaptability of density estimation to different types of datasets and biases. For cumulative distribution function estimation, the package includes the empirical estimator proposed by @cox2005 and the kernel-type estimator by @bose2022, along with several bandwidth selectors for the latter. Finally, the package includes @muttlak1988 real length-biased dataset on shrub width as an example dataset.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/noeliasanchmrt/WData}
  \item Report bugs at \url{https://github.com/noeliasanchmrt/WData/issues}
}

}
\author{
\strong{Maintainer}: Sánchez Martínez Noelia \email{noelia.sanchez.martinez@rai.usc.es} (\href{https://orcid.org/0000-0002-2869-1347}{ORCID})

Authors:
\itemize{
  \item Borrajo Garcia Maria Isabel \email{mariaisabel.borrajo@usc.es} (\href{https://orcid.org/0000-0002-3372-1993}{ORCID})
  \item Conde Amboage Mercedes \email{mercedes.amboage@usc.es} (\href{https://orcid.org/0000-0003-0306-8142}{ORCID})
}

}
\keyword{`r}
\keyword{internal}
\keyword{lifecycle::badge('experimental')`}
