% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WData-package.R
\docType{package}
\name{WData-package}
\alias{WData}
\alias{WData-package}
\title{WData: Density, Distribution, and Quantile Estimation for Biased Datasets}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Set of tools for analyzing and modeling data that may be subject to biases in sampling. It offers functions to estimate density functions, cumulative distribution functions, and quantiles from biased samples of a continuous distribution. Regarding density function estimation, the package includes Bhattacharyya et al. (1988) and Jones (1991) density estimators and various bandwidth selectors for Jones (1991) density estimator, enhancing the flexibility and adaptability of density estimation to different types of datasets and biases. For cumulative distribution function estimation, the package includes the empirical estimator proposed by Cox (2005) and the kernel-type estimator by Bose and Dutta (2022), along with several bandwidth selectors for the latter. For quantile function estimation, it provides two empirical estimators (one of which corresponds to Sen’s (1984) proposal) and one kernel-type estimator. Additionally, for sparsity function estimation, the package includes the two estimators proposed by Akbari et al. (2019) (one plug-in and one kernel-type), as well as a third kernel-type estimator. Moreover, it offers functions for simulating datasets given a bias function and a reference distribution, providing users with the ability to generate synthetic datasets for various modeling and analysis purposes. It also includes functions for constructing boxplots and QQ plots for biased data, facilitating graphical exploration. Finally, the package includes Muttlak’s (1988) real length-biased dataset on shrub width as an example dataset.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/noeliasanchmrt/WData}
  \item Report bugs at \url{https://github.com/noeliasanchmrt/WData/issues}
}

}
\author{
\strong{Maintainer}: Sánchez Martínez Noelia \email{noelia.sanchez.martinez@rai.usc.es} (\href{https://orcid.org/0000-0002-2869-1347}{ORCID})

Authors:
\itemize{
  \item Borrajo Garcia Maria Isabel \email{mariaisabel.borrajo@usc.es} (\href{https://orcid.org/0000-0002-3372-1993}{ORCID})
  \item Conde Amboage Mercedes \email{mercedes.amboage@usc.es} (\href{https://orcid.org/0000-0003-0306-8142}{ORCID})
}

}
\keyword{`r}
\keyword{internal}
\keyword{lifecycle::badge('experimental')`}
